Revert "Don't run make install with sudo."
If we run the whole build process with sudo chances are that the root
user won't know nothing about unprivileged user's environment variables
resulting in a build failure.  The best thing I can do here is adding
an argument to the "compile" script that let's users indicate which
wrapper to "make install" they are going to use (sudo, nothing at all or
whatever they want).

This reverts commit 3ba38676f284e06cdf965977a4463980abc96dd7.